Solve issues with up/down cursor movements when re-rendering model.
If we have multiple lines corresponding to the same address, then we need
to be able to identify each of them uniquely. So, in model, each line is
identified by tuple (addr, subno), where subno is a relative number of line
for the same addr. With such numbering, there's a great chance of
re-rendering model and still be able to find exact line which was current
with a previous model. For a slight chance of mismatch (scenario: cursor
was positioned on last of few comment lines, "delete comment" action was
executed), line with subno==0 will be used.