curses: Add ucurses from micropython-lib.
curses is a dependency of npyscreen. ucurses is an implementaion of curses
subset in pure Python and in a way allowing to share terminal with non-curses
components easily (like pyedit component already used by ScratchABit). The
directory named "curses" because it's, again, what npyscreen uses. Yes, that
means that it is intended to override "curses" package from CPython stdlib.