scratchabit: Implement a way for CPU plugin to contribute to help.
If CPU plugin class has help_text attribute, its content will be appended
to a general help as shown on F1 press. This is extension to IDAPython
API, because it doesn't seem to provide anything like that. The closest
thing it offers is per-asm_t "help screen id", but that looks like
per assembler syntax variant help, not per-CPU help, and it's unclear how
to define that "help screen".