engine: Rework control flow tracing so first current function was explored.
For this, split analysis stack into call stack vs branch stack, and
process branch stack first. There're of course real-world issues like
tail calls, basic support for pre-detecting those, if possible, is added
Complete detection will require post-analysis. Like, if something (1)
turned out to be a function, but was jumped to initially, and thus was
included in the body of another function (2), those address ranges should
be removed from (2). This works well if (1) was detected to be a function,
i.e. was called. If it never was, then including it in the body of (2)
is right in the first place.