scratchabit: Introduce a concept of subareas.
This can be viewed as another way to partition usual "areas", in addition
to functions. Unlike functions, whose detection is tied to control flow
(or information provided by loader), subareas are completely user-defined,
and in general, orthogonal to division by function. However, a typical
usecase it's intended to solve is subdivision by the original source (or
object) module. In this case, there would be containment hierarchy:
area - subarea - function. (And subareas could be set by a loader, too.)
One requirement for subareas is that they don't overlap.