engine: analisys_stack_push: Implement special handling of return addresses.
Previously, ScratchABit employed "breadth first" strategy for handling
function calls - when processing a function, all code paths "within" it
were processed first, and any sub-calls a function made were added to a
queue. However, not each call does return. In this case, bytes following
the call is not reachable (and doesn't have to be the code), but
ScratchABit tries to decode them as code nonetheless, which may lead to
decoding conflicts.

So, switch to depth-first decoding process, where destination of a call
is processed first, and address following the call instruction (i.e.
expected return address) - only after that. This should lead to more
precise code separation and avoid code vs data/etc. conflicts.

To accommodate this, a new flow flag is added, fl_RET_FROM_CALL, which
is supposed to be returned by CPU plugin instead of fl_F. It's thus an
extension to IDAPython API. Also, add more readable aliases for
IDAPython's fl_CN and fl_JN.

This change as is breaks function boundary discovery and thus needs to
be elaborated further.