engine: Add set_arch_id() and code_addr_mask var.
set_arch_id() can be used to set current used arch/subarch/mode, and
indirectly, value of code_addr_mask variable. The latter can be used
to convert value of the arch-encoded address into real address. For
example, for ARM, LSB == 1 of an address means switching to Thumb mode,
and need to be masked out to get a real instruction address.